Fix show-all-calls bug
Even if show-all-calls was set to False, multiple call lines
would be shown in case of multiple call sites within a
single block. This happened because de-duplication was
not done in the right place and the existing de-duplication
had a bug (wrong variable being used).

This change moves de-duplication in the graph building
code, adds a test for it and rebuilds the examples.