Fix generation of "nested" connections
Handle special cases of nodes that are empty or that have only noop
lines.

As part of this, remove __begin__ if a label is defined on line 1, since this
would lead to multiple nodes with the same line number, which introduces
potential non-determinism in the code that generates nested connections.