Improve EOF pruning logic
In addition to removing the eof pseudo-node, all connections to it need
to be pruned away, otherwise it will still be shown on the graph.

Moreover, EOF can be pruned away only if it's not reachable by nested
and call connections, otherwise it will be executed (if it contains
code).

Oh, how much fun it is to read/write cmd files.